+++ src/screens/TaskRankingScreen.js
@@ -60,6 +60,8 @@ const TaskRankingScreen = () => {
   ]);
 
   const handleDeleteTask = (id) => {
+    const confirm = window.confirm("Do you want to delete this task?");
+    if (!confirm) return;
     dispatch(removeTask(id));
     setTaskDetailsView(false);
   };
