+++ src/screens/TasksListScreen.js
@@ -45,6 +45,8 @@ const TasksListScreen = () => {
   }, [dispatch, userDetails, navigate, deleteTaskMessage, pageNumber]);
 
   const handleDeleteTask = (id) => {
+    const confirm = window.confirm("Do you want to delete this task?");
+    if (!confirm) return;
     dispatch(removeTask(id));
     setTaskDetailsView(false);
   };
